not done

+++ app/controllers/Cart.php
@@ -4,12 +4,16 @@ namespace app\controllers;
 class Cart extends \app\core\Controller{
 	public function index(){
         $cart = new \app\models\Cart();
-		$carts = $cart->getAll();
-		$this->view('Product/index', $carts);
+		$carts = $cart->getForUser([$_SESSION['user_id']]);
+		$this->view('Cart/index', $carts);
 	}
 
     public function delete($cart_id){
         $cart = new \app\models\Cart();
+        if ($cart->user_id != $_SESSION['user_id']) {
+            header('location:/Product/index');
+            return;
+        }
         $cart->deleteAt($cart_id);
         header('location:/Cart/index');
 	}
